refactor: prefix pipeline steps with indexes
