+++ plugin-server/src/worker/ingestion/event-pipeline/3-emitToBufferStep.ts
